Add command line for selecting simple/advanced scenarios for api-demo
