+++ api-demo/assets/simple-input.ts
