+++ api-demo/src/advancedDemo.ts
@@ -1,0 +1,6 @@
+import * as colors from 'colors';
+import * as os from 'os';
+
+export function advancedDemo(): void {
+  console.log(colors.yellow('Demo scenario: advanced') + os.EOL);
+}
