+++ api-demo/src/simpleDemo.ts
@@ -1,0 +1,66 @@
+import * as colors from 'colors';
+import * as fs from 'fs';
+import * as path from 'path';
+import * as os from 'os';
+import { TSDocParser, ParserContext, DocComment } from '@microsoft/tsdoc';
+import { Formatter } from './Formatter';
+
+/**
+ * The simple demo does not rely on the TypeScript compiler.
+ * It uses the default parser configuration.
+ */
+export function simpleDemo(): void {
+  console.log(colors.yellow('Demo scenario: simple') + os.EOL);
+
+  const inputFilename: string = path.resolve(path.join(__dirname, '..', 'assets', 'simple-input.ts'));
+  console.log('Reading assets/simple-input.ts...');
+
+  const inputBuffer: string = fs.readFileSync(inputFilename).toString();
+
+  // NOTE: Optionally, can provide a TSDocParserConfiguration here
+  const tsdocParser: TSDocParser = new TSDocParser();
+  const parserContext: ParserContext = tsdocParser.parseString(inputBuffer);
+
+  console.log(os.EOL + colors.green('Input Buffer:') + os.EOL);
+  console.log(colors.gray('<<<<<<'));
+  console.log(inputBuffer);
+  console.log(colors.gray('>>>>>>'));
+
+  console.log(os.EOL + colors.green('Extracted Lines:') + os.EOL);
+  console.log(JSON.stringify(parserContext.lines.map(x => x.toString()), undefined, '  '));
+
+  console.log(os.EOL + colors.green('Parser Log Messages:') + os.EOL);
+
+  if (parserContext.log.messages.length === 0) {
+    console.log('No errors or warnings.');
+  } else {
+    for (const message of parserContext.log.messages.map(x => x.toString())) {
+      console.log(inputFilename + message);
+    }
+  }
+
+  console.log(os.EOL + colors.green('DocComment parts:') + os.EOL);
+
+  const docComment: DocComment = parserContext.docComment;
+
+  console.log(colors.yellow('Summary: ')
+    + JSON.stringify(Formatter.renderDocNode(docComment.summarySection)));
+
+  if (docComment.remarksBlock) {
+    console.log(colors.yellow('Remarks: ')
+    + JSON.stringify(Formatter.renderDocNodes(docComment.remarksBlock.nodes)));
+  }
+
+  for (const paramBlock of docComment.paramBlocks) {
+    console.log(colors.yellow(`Parameter "${paramBlock.parameterName}": `)
+    + JSON.stringify(Formatter.renderDocNodes(paramBlock.nodes)));
+  }
+
+  if (docComment.returnsBlock) {
+    console.log(colors.yellow('Returns: ')
+    + JSON.stringify(Formatter.renderDocNodes(docComment.returnsBlock.nodes)));
+  }
+
+  console.log(colors.yellow('Modifiers: ')
+    + docComment.modifierTagSet.nodes.map(x => x.tagName).join(', '));
+}
